backend_e_frontend/requirements.txt
streamlit
azure-storage-blob
azure-cosmos
pydub
pandas